+++ game-rest/src/main/java/com/kevinguanchedarias/owgejava/configuration/RestConfiguration.java
@@ -2,6 +2,7 @@ package com.kevinguanchedarias.owgejava.configuration;
 
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.converter.HttpMessageConverter;
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
@@ -11,7 +12,7 @@ import java.util.List;
 
 @Configuration
 public class RestConfiguration extends WebMvcConfigurationSupport {
-    
+
     @SuppressWarnings("rawtypes")
     @Override
     protected void extendMessageConverters(List<HttpMessageConverter<?>> converters) {
@@ -20,6 +21,7 @@ public class RestConfiguration extends WebMvcConfigurationSupport {
                 ObjectMapper mapper = ((MappingJackson2HttpMessageConverter) converter).getObjectMapper();
                 mapper.setSerializationInclusion(Include.NON_NULL);
                 mapper.setDefaultPropertyInclusion(Include.NON_NULL);
+                mapper.configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, false);
             }
         }
     }